# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "fsGaze API",
    "description": "API for analyzing KerML documents and managing safety analysis data.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/safety/failure-modes": {
      "post": {
        "tags": ["Failure Modes"],
        "summary": "Create a new failure mode",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFailureMode"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Failure mode created" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/safety/failure-modes/{uuid}": {
      "get": {
        "tags": ["Failure Modes"],
        "summary": "Get a failure mode by UUID",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful retrieval" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Failure Modes"],
        "summary": "Update a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFailureMode"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Update successful" },
          "400": { "description": "Bad request" }
        }
      },
      "delete": {
        "tags": ["Failure Modes"],
        "summary": "Delete a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deletion successful" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/safety/failure-modes/{uuid}/causes": {
      "get": {
        "tags": ["Causation"],
        "summary": "Get causes for a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful retrieval" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/safety/failure-modes/{uuid}/effects": {
      "get": {
        "tags": ["Causation"],
        "summary": "Get effects of a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful retrieval" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/safety/causations": {
      "post": {
        "tags": ["Causation"],
        "summary": "Create a causation link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCausation"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Causation created" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/safety/causations/{uuid}": {
      "delete": {
        "tags": ["Causation"],
        "summary": "Delete a causation link",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deletion successful" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/safety/failure-modes/{uuid}/risk-ratings": {
      "get": {
        "tags": ["Risk Ratings"],
        "summary": "Get risk ratings for a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful retrieval" },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "tags": ["Risk Ratings"],
        "summary": "Create a risk rating for a failure mode",
        "parameters": [
          { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRiskRating"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Risk rating created" },
          "400": { "description": "Bad request" }
        }
      }
    },
...
    "/api/safety/failure-modes/{uuid}/safety-requirements": {
        "get": {
            "tags": ["Safety Requirements"],
            "summary": "Get safety requirements for a failure mode",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "post": {
            "tags": ["Safety Requirements"],
            "summary": "Create a safety requirement for a failure mode",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyRequirement"
                        }
                    }
                }
            },
            "responses": {
                "201": { "description": "Safety requirement created" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/failure-modes/{uuid}/notes": {
        "get": {
            "tags": ["Safety Notes"],
            "summary": "Get notes for a failure mode",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        },
        "post": {
            "tags": ["Safety Notes"],
            "summary": "Create a note for a failure mode",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateSafetyNote"
                        }
                    }
                }
            },
            "responses": {
                "201": { "description": "Note created" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/ports/{uuid}/failure-modes": {
        "get": {
            "tags": ["Failure Modes"],
            "summary": "Get failure modes for a port",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        }
    }
  },
  "components": {
    "schemas": {
      "CreateFailureMode": {
        "type": "object",
        "properties": {
          "existingElementUuid": { 
            "type": "string",
            "description": "UUID of the existing element (SW component or port) to link the failure to",
            "example": "12345678-1234-5678-9012-123456789012"
          },
          "failureModeName": { 
            "type": "string",
            "description": "Name of the failure mode",
            "example": "Communication Loss"
          },
          "failureModeDescription": { 
            "type": "string",
            "description": "Detailed description of the failure mode",
            "example": "Loss of communication between components due to network failure"
          },
          "asil": { 
            "type": "string",
            "description": "ASIL (Automotive Safety Integrity Level) rating",
            "enum": ["QM", "A", "B", "C", "D"],
            "example": "C"
          }
        },
        "required": ["existingElementUuid", "failureModeName", "failureModeDescription", "asil"]
      },
      "UpdateFailureMode": {
        "type": "object",
        "properties": {
          "failureModeName": { 
            "type": "string",
            "description": "Updated name of the failure mode"
          },
          "failureModeDescription": { 
            "type": "string",
            "description": "Updated description of the failure mode"
          },
          "asil": { 
            "type": "string",
            "description": "Updated ASIL rating",
            "enum": ["QM", "A", "B", "C", "D"]
          }
        },
        "required": ["failureModeName", "failureModeDescription", "asil"]
      },
      "CreateCausation": {
        "type": "object",
        "properties": {
            "sourceFailureModeUuid": { 
              "type": "string",
              "description": "UUID of the failure mode that acts as the cause",
              "example": "12345678-1234-5678-9012-123456789012"
            },
            "targetFailureModeUuid": { 
              "type": "string",
              "description": "UUID of the failure mode that acts as the effect",
              "example": "87654321-4321-8765-2109-876543210987"
            }
        },
        "required": ["sourceFailureModeUuid", "targetFailureModeUuid"]
      },
      "CreateRiskRating": {
          "type": "object",
          "properties": {
              "severity": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Severity rating (1-10)",
                "example": 7
              },
              "occurrence": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Occurrence rating (1-10)",
                "example": 3
              },
              "detection": { 
                "type": "integer",
                "minimum": 1,
                "maximum": 10,
                "description": "Detection rating (1-10)",
                "example": 5
              },
              "ratingComment": { 
                "type": "string",
                "description": "Optional comment about the risk rating",
                "example": "Based on historical data and expert judgment"
              }
          },
          "required": ["severity", "occurrence", "detection"]
      },
      "CreateSafetyRequirement": {
          "type": "object",
          "properties": {
              "name": { 
                "type": "string",
                "description": "Name of the safety requirement",
                "example": "Emergency Stop Function"
              },
              "reqID": { 
                "type": "string",
                "description": "Unique identifier for the requirement",
                "example": "REQ-001"
              },
              "reqText": { 
                "type": "string",
                "description": "Detailed text of the requirement",
                "example": "The system shall provide an emergency stop function that can halt all operations within 100ms"
              },
              "reqASIL": { 
                "type": "string",
                "description": "ASIL level for the requirement",
                "enum": ["QM", "A", "B", "C", "D"],
                "example": "D"
              },
              "reqLinkedTo": {
                "type": "string",
                "description": "Optional reference to external requirement",
                "example": "ISO-26262-3:2018"
              }
          },
          "required": ["name", "reqID", "reqText", "reqASIL"]
      },
...
      "CreateSafetyNote": {
          "type": "object",
          "properties": {
              "note": { 
                "type": "string",
                "description": "Content of the safety note",
                "example": "This failure mode requires special attention during testing phase"
              }
          },
          "required": ["note"]
      }
    }
  }
}